+++ src/entities/inline/queryResult/InlineQueryResultLocation.php
@@ -14,8 +14,11 @@ class InlineQueryResultLocation extends InlineQueryResult
     public float $latitude;
     public float $longitude;
     public string $title;
-    public int $live_period;
-    public string $thumb_url;
-    public int $thumb_width;
-    public int $thumb_height;
+    public ?float $horizontal_accuracy;
+    public ?int $live_period;
+    public ?int $heading;
+    public ?int $proximity_alert_radius;
+    public ?string $thumb_url;
+    public ?int $thumb_width;
+    public ?int $thumb_height;
 }