+++ src/entities/message/Location.php
@@ -14,4 +14,8 @@ class Location extends BaseObject
 {
     public float $longitude;
     public float $latitude;
+    public ?float $horizontal_accuracy;
+    public ?int $live_period;
+    public ?int $heading;
+    public ?int $proximity_alert_radius;
 }